+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 
+ * Upgrade dependencies for OpenCV 4.12.0, FFmpeg 8.0
+
 ### June 30, 2025 version 1.5.12
  * Add `LibgdxFrameConverter` to convert `Frame` to Libgdx `Pixmap` ([pull #2315](https://github.com/bytedeco/javacv/pull/2315))
  * Fix `FFmpegFrameRecorder` dropped frame issues with audio samples ([pull #2307](https://github.com/bytedeco/javacv/pull/2307))